requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
astunparse==1.6.3
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
blinker==1.6.2
boto3==1.28.3
botocore==1.31.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
databricks-cli==0.17.7
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.98.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.4.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.0.0
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kafka-python==2.0.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
minio==7.1.15
ml-dtypes==0.2.0
mlflow==2.5.0
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.3
psycopg2-binary==2.9.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
waitress==2.1.2
websocket-client==1.6.0
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
